re #31
The big code clean up:
 
The whole kernel build leaks memory and double deletes:
Valgrind tells us that 800 problems exist and about 300 are in 
our code. This is an attempt to clean up this mess.

Algorithm is the first to be cleaned. Errors on double deletion of 
a vector where found [finalize]. Checking for empty strings found , and failure
to delete managed memory when exiting  on a fail [execute].

Access to subalgorithm has been changed to hand back a 
std::vector<Algorithm*>& 
This has both a const / non-const form.