re #57
   Modifications of a multi-part submission to TripleIterator.
This addition added the non-const operator* and operator-> to 
TripleIterator. The std iterators have these two operators as
const for const_iterator and non-const for iterator. So this iterator 
could/should be split. 

Addition of the first of several tests, using std::alogrithm functions
that used this iterator. In this case , std::copy.