re #53
Updated LoadRaw algorithm to read the SpectraDetectorMap. This is implemented as a sub-algorithm calling the LoadMappingTable. There is a slight penalty in doing that since the raw file is open another time but it is minimal compared to the time it takes to read data. 
The test LoadRawTest has been modified accordingly but there is currently a problem in the HET_definition.xml which does not contain all the individual pixel geometric information. New code in the LoadRawtest.h file is working (test with reading instrument from raw file instead of xml) but has been commented out for the moment.