re #531 Changes to Properties files
Mantid now tries to load TWO properties files
Mantid.Properties as before
and now a Mantid.user.properties

Any entries in the user.properties file overides those on the main file.  The intention is that on reinstalls we will not overwrite the user file, but are free to replace the original .properties file to add new keys etc.

IMPORTANT:
The release properties files are now in the Properties subdirectory, the ones in the release directory have been removed.