Run same test using Matlab's default solver (which does not use derivative) on Gaussian tests and this solver does not do any better than the GSL solvers.
The verdict at the moment is that the GSL simplex is more robust than the GSL Levenberg-Marquardt, and constraints appear to have the most beneficial effect when used together with the simplex. 

When bad initial values are used with the GSL Levenberg-Marquardt it may refuse to do anything. For this scenario code has been put in place which then tried to use the simplex method. Refs #1044.