re #1611 - indirectAnalysis adjustments based on comments from Spencer Howells
Fury tab - added "Plot" checkbox which makes use of the the plotFury function in the IndirectDataAnalysis python module. Also adjusted this function so that the cut off point for the x-axis is based on when the y values are successively zero, rather than when they go above one.

Slice tab - added "Plot Raw" button and suitable function. Adjusted the calibration input so it is optional based on a new checkbox. Adjusted the second TOF range so that it is optional based on a new checkbox. Cleaned up the function generally to remove unnecessary relics before the distinction was clear between slice/elwin. Made multiple input files an option.