Extend LoadInstrument to include the input properties:
  InstrumentName

This means there will be two input properties which can be used to specify and IDF:

  Filename
  InstrumentName

If Filename is specified then use that this filename. If not then it is required that an InstrumentName is specified. And an appropriate filename is created at runtime (search the Workspace's Run object for "startdate"/"starttime" and get IDF search directory from Config) 

Modify all loaders (except LoadEmptyInstrument) to load inst using this InstrumentName property
 
re #1997