re #2239
Changes to Algorithm to allow subAlgorithms to be cancelled.

Now algorithms: 
Maintain a vector of weak pointers to any subalgorithms they create. Weak pointers will not prevent the algorithms from being destructed. 
On receiving a cancel command will pass it on to any valid sub algorithms. 
Delete the list of weak pointers when the algorithm is destroyed. As they are weak pointers they have no effect of the lifetime of and subalgorithms.