1) Now only convertValue() unit method in IFunction. 2) Spotted a missing result-unit="TOF^2 for bank one in GEM IDF 3) Removed warning message that IkedaCarpenter can only be used with TOF 4) Added IkedaCarpenterPVTest back in
Added code that: one way to speed up IkedaCarpenter further would be:

      // note if a version of convertValue was added which allows a double* as first argument
      // then could avoid some copying above plus only have to resize m_wavelength when 
      // its size smaller than nData

But this has to be for another day.

When testing this function consider using focussed GEM data (e.g. test/nexus/focussedGEM38370_TOF.nxs) since at present it is only
for the GEM instrument that IkedaCarpenter parameter are currently pre-set.

re #1547.