re#1839 - The problem was commands like Ctrl+C on script interpreter is routed to global editing command handlers in Application window. Added code to handle Ctrl+C on script widget in the application window copySelection() method.
Application window copySelection() handles copy actions on all windows of type MDISubWindow in Mantidplot and 
if any of these windows is kept opened and a Ctrl+C on script interpreter widget the command was getting routed to application window code.
Fixed this.However when I do paste command (Ctrl+V shortcut ) on script  interpreter it gives an error -conflicting
Ctrl+V shortcut Need to fix this.