When workspace not set default wavelength in IkedaCarpenter (IC) to one.
Introduced a number places where if the workspace or instrument of the workspace is
not set properly that warnings are displayed.

Removed now obsolete constant wavelength option in IC code. 

Re-enable the IC test and further improve it. 

Note workspace of fitting function will normally automatically
be set. But for now one exception is when using 

  IFitFunction *out = FunctionFactory::Instance().createInitialized(alg2.getPropertyValue("Function"));
  IPeakFunction *pk = dynamic_cast<IPeakFunction *>(out);  

To set the workspace for pk add a line like:

  pk->setMatrixWorkspace(ws2D, 0,-1,-1);

re #2446