Next step in overlay capability. This refs #3958.
In order to make it easier for the overlay capability to happen,
I had to make the plugin mode destroy the current sources when
then view window is hidden. This required an eventfilter due to
the free floating nature of the view window. For the plugin mode,
this change opens up overlay capability. Now the individual view
need to handle it.