refs #3763 Crude implementation of Dynamic properties.
Will not work properly form framework when property modifier change validator in such a way, that old value of the dependent property is incorrect. Substantial changes needed to rectify that and to avoid such situation, so currently has to be dealt with by user.

It would be also beneficial to implement proper notification mechanism, which would inform the property about the changes in the properties it depends on, as current implementation can potentially have substantial overhead, which user should be really carefull about.