Refs #4121. Removing MAKE_VATES guard for right-click menu.
I am removing the guards for the right-click workspace menu so that the
VATES simple interface can be used with the separate Linux RPMs. This,
however, means that MantidPlot will crash if the VATES libraries do not
exist and the simple interface is called. I am working with Martyn on a
solution. Also, I'm not sure what this will do to the Windows install kits.