Dealing with SaveIsawPeaks for WISH instrument re #4104
Made SaveIsawPeaks quit with an error if a detector module (other than "None") is found that is not of form bankN. In the case of WISH, the detector modules are of form tubeNNN and are not rectangular detectors.

The issue of saving peaks unsuitable for IsawPeaks will be dealt with in another ticket (also critical).

Signed-off-by: Karl Palmen <karl.palmen@stfc.ac.uk>