Add ability to set a curve's symbol, pen & brush via python.
QwtSymbol is exposed as PlotSymbol and allows all aspects of the symbol
to be customised.
The setCurvePen method acts on the characteristics of the line and offers
more advanced options than the setCurveLineColor/Style/Width methods.
setCurveBrush acts on the area beneath a curve (e.g. fill colour, style...)
Re #2656.