Split up mantidplot.py and move it into a directory. Re #1037.
mantidplot.py is starting to get big, and once all the proxy business
for qti objects is added is will really swell. So the proxy objects
will go in a separate file and everything is in a mantidplot directory
that is imported. It shouldn't matter if an old mantidplot.py is lying
around in the old place - python will prefer to import the directory.
Fingers crossed that I've got the installation settings right.