Temp fix for sorting Muon Analaysis ws groups.  Refs #3012.
The source of the problem seems to be that the ws group
"MuonAnalysisGroup" returns no history via its "getHistory" method.
Implemented a temporary measure of presenting the user with a dialog as a
warning to say that one of the workspaces has no history, and so it may not
have been sorted correctly. Raised the ticket #4439 for next iteration.