Refs #4433: folder switching is not correct with floating windows
I get some windows remaining (e.g. Graph2) if it was floating. It is not quite the same all the time, sometimes the window his hidden sometimes not. Frequently I get an annoying flicker of the windows going to front and then to back etc. until I click on a menu bar. Ubuntu 11.04