Expose MdiSubWindow hide() & show() to python. Re #4282.
This makes calling show & hide work from python instead of leaving the
'wrapper' window there.
HOWEVER, I'm not sure we should be overriding these methods - they are not
virtual methods (neither is close()). setVisible() - which these two call
- IS virtual, but when I tried to implement that instead it completely
messed up the appearance of the instrument window and the custom interfaces
for some reason.