Refs #4520 - Reordering of .ext vector before findRun is called.
Change made to FileFinder, to ensure that non-wildcard extensions are
searched for, before extensions like .N*, etc.  This has resulted in a VERY
large improvement in responsiveness of the Convert to Interface tab when
loading files from a large directory across a network.

Also made a (perhaps temporary) change to WorkspaceGroup to deal with a case
where m_wsNames.end() was being called in .contains(), and a fatal error
occuring as a result.  I have yet to find why such an error could be thrown in the first place.