Refs #4622 - Change multi file setting of a MWRunFiles widget.
Previously this was set to true by default, but since default behaviour
has been switched to false, multi file support was lost for the "Run
Files" field of the Energy Tranfer tab of the CTE interface.  Changed the
multi file property to true for this widget so multiple files can be run.