refs #4401 MDEventWSWrapper made reenterable
As boost function pointers do funny things when stings supplied as function arguments and when the number of arguments is high, custom vftable has been build with functions, which use 0 or 1 argument. As the result, internal memory for the class was needed.

Reverting to classic old function pointers allows to use multiple function arguments and eliminate the need for internal class memory. This allows to deploy const methods and makes class reenterable (mainly)