Refs #1419 - Support loading of incorrectly padded runs.
In the case where a user wants to load a file that had an "inccorrect" zero
padding according to Facilities.xml, he/she should be allowed to if that
file exists!

If the user only specifies a single file, then we should use the string they
provided in its untampered form.

Note, files with incorrect zero padding will not work with multi file
loading in its present form.