Refs #1419 - Changing a log message priority from error to debug.
In the case where a user wishes to load a file with a name which is not of the
"standard" format (including a valid instrument), our call to the
ConfigService getInstrument method should not report an error.

A debug message is sufficient, and there is still an exception thrown which
a class user can then interpret in any way he/she sees fit.