Refs #4859 - Remove wildcard extensions before finding runs.
This is the first step of this ticket, which removes the (currently buggy)
findRun function in the case where there are two similar files with
extensions of .s01 and .s02 and .s* has been passed as an extension to look
for.  In that case, only a single (arbitrarily picked) file is returned.

More work will be needed to accomodate extensions with wild cards properly.