Refs #4513 - Initial Raw and RawB translation.
Added a python algorithm which is the first attempt at translating the Raw
and RawB Fortran code.  Note:

1. Backward foil periods are currently assumed to always be the same for
each run.  Check that this is the case.  If not, check that they are always
the same for detectors in the same module.

2. Check that periods used for monitor calculations can be different from
the periods used for corresponding spectra.

3. Is dividing through by a "Time" ws not equivalent to a Rebin?

4. Backward results do not currently match RawB output.  Check that
RAWB.FOR;8 is actually the same code that has been compiled to create the
RawB program currently on the VMS systems.

5. Not sure if runs with less / more than 6 periods should be supported.  If
so, what are the period schemes in those cases?

Will now pass the script and pose questions to Jerry.  Will also make a
start dechipering TFitD.