Refs #5036 - MWRunFiles now finds files asynchronously.
A thread started to find files whenever a user changes the content of the
field and then changes focus to elsewhere.  Both conditions have to be met
before the thread starts.  If the thread is already running, and new input
is detected, then it is cancelled and restarted with the new files to find.