Refs #5036 - Fix for Muon interface - additional MWRunFiles signal.
Added an extra signal to MWRunFiles to help distinguish between when a user
has finished typing (fileEditingFinished()) and when the file finding
process has finished (fileFindingFinished()).  Note that this signal is
different to the filesFound signal in that fileFindingFinished() is emitted
whether the files have been found or not.

Changed the Muon Analysis interface to wait for fileFindingFinished()
instead of fileEditingFinished().