Added double buffer to ImagePlotItem.
Now use a double buffer for the DataArray passed into the
ImagePlotItem, and have ImagePlotItem take ownership of
the DataArray.  Since ImagePlotItem will only delete a
DataArray when it no longer needs it, this should avoid
any possible problems with the system using a "stale"
pointer to draw the image in response to window events.
refs #5058