Re #4960 Updates to Tobyfit resolution methods
Basic simulation of resolution function is now possible for a foreground function
such as CobaltSpinWaveDSHO. However the input data, including chopper and moderator
parameters and detector information is read from a RunParam object. Much of this data
should be extracted from the Instrument, the Sample and log files, though may be stored
in the RunParam objects.
Simple tests are consistent with Tobyfit results. A bug exists in taking account of sample
shape.