Use generateHistogram() to safely get Y-values
Added generateHistogram() to IEventList, and use it to get
a clean copy of the Y-values, rebinned according to the
required X-values.  This does NOT modify the underlying data
as setting an X-scale on the spectra did.  It also seems to
have fixed a couple of rare glitches that may have been
related to the MRU list.
refs #5058