No vertical scroll image now subsamples spectra
If the vertical scroll bar is turned off, the ImageViewer
now displays an image obtained by selecting a subset of
the available spectra.  Specifically, if there are 5,000
spectra and 500 rows available in the image, the image is
formed using every 10th spectrum.  The results can be
misleading due to aliasing, but at least it is possible
to see if there are counts in the data.
refs #5058