Refs #5207 - Disable run button while finding files in C2E.
Added an extra signal in MWRunFiles to make this possible - "FindingFiles".

The run button is now disabled when a user starts to change the files to be
loaded, and enabled when the file finder has finished.  Its text also
changes to "Editing..." and "Finding files..." appropriately.

The MWRunFiles widget itself is disabled when the FileFinder is running.
This is a quick way to solve the problem of a user spamming the field with
multiple requests, and spawning multiple - and potentially computationally
intensive - background file finder threads.  I think a change to how
FileFinder is called within the widget is needed to better enable the thread
to stop and a new one to take over.