Re: 5285 Add convertToBinBoundary() to VectorHelper
Since VectorHelper::rebin() requires current X vector with
one more entry than the current Y vector, even if the
Workspace2D is a distribution, we need to adjust the current
X vector before calling rebin.  This is a convenience method
that essentially reverses the convertToBinCentre() method, so
it seems to belong in VectorHelper.
refs #5285