RE: 5058 Add ArrayDataSource to easily show an array
The ArrayDataSource wraps a simple array of doubles stored
in row major order, so the array can be viewed using the
ImageViewer. This is a convenience class that should make
it easy to apply the ImageViewer in other ways.
Also, modified ImageViewDemo to use the new ArrayDataSource,
as a demonstration of the new class.
refs #5058