Re: #5058 Fix possible error in value readout for vertical graph
Previously, both the x and y coordinates of a pointed at location
on the graphs were both restricted to have valid data source
coordinates.  However, that was a mistake.  For the horizontal
graph, only the x-coordinate should be restricted and for the
vertical graph, only the y-coordinate should be restricted.
refs #5058