Refs #5441 - *Added* multi file functionality to Load dialog.
Now allowing devs to ask that MWRunFiles deals with file finding via a
specified algorithm property instead of the default method of passing the
user's inputs straight to FileFinder, which is ill-equipped to deal with
complex multi file loading / addition, etc.

LoadDialog now uses MWRunFiles in this way, via the "Filename" property
of Load.

This indirectly solves most (all?) of the recent problems surrounding the
Load dialog, which are a result of my lack of understanding in how it could
be affected by adding the multi file functionality.  I had certainly not
intended for LoadDialog to be even used for multi-file loading at this
stage, and so had simply not catered for it or spotted the problems caused.
Mea culpa.

There are still some outstanding issues:

1) Currently, the suggested ws name given to Load by the dialog is just the
"base" of the first filename, i.e. loading IRS21000+21009.raw produces a ws
with the name "IRS21000".  This name is neither helpful or accurate.

2) After loading multiple files and then reopening the dialog, the
MWRunFiles widget contains the full file name of the first file ONLY.  This
issue is compounded by the fact that specifying addition / ranges of files
during multi file loading is supported via the syntax

[full_path][inst][run]+[run][.ext]

BUT NOT

[full_path][inst][run]+[full_path][inst][run][.ext].

As far as I can tell it is only going to be possible to either:

a) display the full file path of the files that have been found but not
specify which files were added to which; OR

b) accurately display which files are being added to which, but not display
their full files path.

If b) is to be implemented I can not comment on how big a loss (if at all)
not seeing full file paths would be.  a) Does not seem to be a reasonable
option.