Re #5455. Change Events.h comments wrongly referring to nanoseconds.
Also act on a TODO about changing the TofEvent member variable visibility
back to protected from public. This accounts for most of the changes
as people had just used the members directly instead of accessor methods.
A few places were using it to modify the members directly. I was able to
change the logic to get around this in all cases except for LoadEventNexus.