Refs #5411 - Better handling of F2Py imports for Indirect scripts.
Check for numpy package version as well as OS.  Highlighted out win64 and
lnx64 parts for now.  Should just be able to uncomment them, and add the
libraries when they are ready to be added.

Also removed the annoying error messages popping up during system tests.

Still a stop-gap measure until compilation of the Fortran code can be put
into the automated build process of Mantid.