Re 5533: Comment out OMP Pragma
If the OMP pragma is included in IntegratePeaksMD, the algorithm seg faults
sporadically when processing multiple TOPAZ runs in a script, on
Scientific Linux 6.2.  Typically, it seg faults after 2 to 6 runs are
processed, though occasionally it will process all 8 requested in the
script without crashing.  Since the lower level codes already use OpenMP,
parallelizing at this level is only marginally useful, giving about a
5-10% speedup.  Perhaps it should just be removed permanantly, but for
now it is commented out to avoid the seg faults.
Refs #5533