Re 5585, IndexPeaks now indexes using per run optimzied UB
When indexing peaks in a peaks workspace, the UB matrix now
is optimized for the list of peaks from each run, before
actually using it to index the peaks in that run.  This will
increase the number of peaks indexed, in particular when the
peaks workspace contains peaks from multiple runs.
refs #5585