Re #5570. First cut at an algorithm to combine peaks workspaces.
Follows exactly what's written in the ticket. Peaks in the second
workspace that are within the given Q tolerance in all directions
don't make it into the output.
Note that it's quite possible for a given peak in the first workspace
to be matched to multiple peaks in the second, if the tolerance
is large enough/the peaks are close together.