Refs #5212, #5441 - Better parsing when multifile loading. Hist Fixed.
Allow filenames specified with the format <<short><operator>>...<short>
where:

<short> = <dir><inst><underscore><runs><ext>
<operator> = "+" or ","
<runs> = lists and/or ranges of (possibly added) runs e.g. "12,13+14,15:20".

See header file of MultipleFileProperty for more info on syntax.

This enables workspaces loaded through the algorithm to have a history that
can be used to reproduce the workspace properly.

MWRunFiles widget of LoadDialog now remembers the exact text a user typed -
in this way any user who types "INST0-50", clicks "Run" and reopens the
dialog is not met by a huge string containing a list of all 51 fully
resolved files, but rather their original input.

Changed the vector<vector<string>> templated versions of the toString and
toValue functions to accept customised delimiters.  This enables us to add
an option for the user to turn off multifile parsing if they are crazy
enough to want to load files with a plus or a comma in their path.

Changed the inner workings of Load for multi files.  No longer calling sub
algorithms with setChild or alwaysStoreInADS.  All workspaces are
essentially created outside the ADS and managed by the algorithm, right up
until the OutputWorkspaces are set.

Tests added where appropriate, but the majority of parsing cases are already
covered by MultiFileNameParser.