Refs #5665 - Remove Y-Axis values when autoscale is selected.
Remove (and store) the contents of the Y-Axis fields when the autoscale
checkbox is checked, replacing them with "N\A".  Also, disable the minimum
field when "Start at First Good Data" is chosen.

These two changes now ensure a bigger visual difference between the concept
of "disable and do not use", and "disable but still use", and clears the way
for adding a "Custom Value" option to the dropdown box.