Several changes to job status dialog (and job submission in general)
1) HTTPS is now enabled and enforced.  No unencrypted connections are
allowed.  (Note: We're accepting the certificate from the server as-is
with no validation.  That's not really correct, but it's the best we can
do for now.)
2) Jobs in the job status dialog are now sorted by their ID's
3) Had to disable the use of QTemporaryFile objects for downloading the
output files because the Load algorithm runs in a separate thread and
the object was cleaning up its file before the algorithm had a chance
to run.

Refs #4884