LoadISISNexus2 gets instrument parameters from Nexus file re #5443
Instrument parameters are got the old way (from IDF) if not found in Nexus file.

Default version was set to 1 to get this to work. It needs setting back to 0, so that old NexusProcessed files can be read. For this code for LoadISISNexus2 to communicated a default version will be added.

Also LoadISISNexus2 loads instrument parameters a second time for an additional log. This is still done the old way.