Properly handle redundant device descriptor packets
There are certain cases when SMS will send device descriptor packets
more than once per run.  Need to ignore the descriptions for devices we
already have (because Mantid will throw an exception if you try to add a
property to a workspace that already has one of the same name).

Refs #5857