Re #5883. Use a different user properties file for the MPI build.
The fermi cluster at ORNL uses the usual SNS home areas (so the same
as the analysis workstations, for example). If the user properties
file is the same, confusion could arise if things are set in it
in and interactive MantidPlot session.
The MPI build now uses Mantid-mpi.user.properties instead.