Re #5883. Logging changes for MPI build.
The MPI build no longer logs to file by default (as this duplicates
the console channel which ends up in a log file for a job).

The format of log messages to the console has been changed to include
hostname & process id so that messages can be separated
according to which mpi rank they relate to. The rank is explicitly
logged at startup.