refs #5873 this looks like fix it
Disabled the write cache for input workspaces and SliceMD/BinMD algorithms. Now it generates(slices) approximately the same number of events as Horace at approximately the same amount of time. This amount is still different from Horace with similar setting but not fundamentally different, so this is the topic of another ticket.