refs #6091 CreateMDWorkspace.cpp supports _PreprpocDetectors parameter
in more or less consistent way,e.g when UsePreprocessedDetectors switch is selected, CreateMDWorkspace.cpp uses detectors in _PreprpocDetectors workspace, which is created on first run on the first file.

when this parameter is not selected, no visible PreprocessedDetectors ws is added to dataservice and algorithm calculates preprocessed detectors each time it is run.