Re #6152. Fix if-else construct that was identical on either branch.
There's also a slight change in behaviour. If a 'good' region contains
more than one log value, and log values are considered to represent the
left of a bin, then the end time will be the point it went 'bad'
(previously it was the time of the last good value + a tolerance).