refs #6220. Fixed. Can now load in PV File Backed.
The prolem actually occured with the new reader. When the PV framework was
deciding which reader to user (running canReadFile) it passed over the
vtkNexusPeaksReader. The canReadFile implementation here opened the nexus
file, but did not close it. Thefore it could not be reopened when the
correct reader was finally chosen by ParaView.