refs #6272 Auto refresh introduced.
An extra button and an extra menu option has been added for this. There are two points that I want to address and check. Firstly, we currently have a lock mode, which is a kind of auto update, but only for when the sliders are moved. Its probably confusing to have two forms of autoupdate, so we should probably roll them into the single lock mode. This should be easy to do. Secondly, if we do not opt for this, it might be best to make the auto refesh icon look something like the current refresh icon, to save confusion.