Re #6270. Improve precision of run_start/end workspace log values.
Previously the precision of these would be the nearest second as
the setting went via the DateAndTime::toISO8601String() method,
which truncated the time even if it had a better precision than
that (as the SNS files do). This then caused some oddities when
viewing logs because they do preserve the incoming precision.

DateAndTime::toISO8601String() now preserves the nanosecond precision
of the underlying DateAndTime class.