Re #6308. Add a .gitattributes file setting the new 'text' attribute
From git 1.7.2, there's a new way of dealing with line endings.
This setting of the attribute should mean that developers get their
native line endings locally, but that we only get LF endings in the repo.
Requiring everyone to set a config property to enforce this should
no longer be necessary.