refs #6311. Query method and other optimsiations.
Some work has been done on caching calculated values and therefore the time for the setSlicePoint has been reduced by about 5%. I added a public method called isViewable, which should allow the client code to pre-query the PhysicalPeak before painting the QtWidget (expensive). While this wont show up in the peformance tests it should make a big difference to the actual usage peformance because usually only a small number of the total available peaks are interacting with the slice at any time. Overall, the changes made have produced a slight imporvement in the measureable performance.