Re #6149. Major changes to RefDetectorViewer to eliminate duplication
...of code w.r.t. the ImageViewer.

The small extensions in RefImageDisplay & RefImagePlotItem compared
to their ImageViewer counterparts have been isolated and these
classes now derive from the ImageViewer ones.

In order to disentangle some things a new class 'RefLimitsHandler'
has been introduced to handle the getting and setting of the values
in the top-right line edits (Peak/Back/TOF Left/Right/min/max).
This is quite a lot simpler than the previous way as the values are
only held by the QLineEdit objects and not separately in addition.