Re #6337. Change return type of (I)TableWorkspace::addColumn
...from a success/failure boolean to a shared pointer to the new column.
Saves repeatedly having to go and fetch the column separately in an
algorithm I'm working on. Any existing places that check the return value
(there aren't many) will still work as the pointer will convert to bool.

I kept the return type of the exposed python method unchanged as the
Column class is not exposed to python and I didn't see any reason to do
so just for this.