BugSolved FindBeamCenter wrong answer
SANSRunWindow.cpp : add the centre information from the Geometry tab,
so to make it to be considered in SingleMode.
ISISCommandInterface.py: original does not take the beam_center position,
but get the position of the center of the bank detector. Avoiding the
wrong assumption that the detector bank will always be centered at its
defined beam_center.
isis_instrument.py : add a method cur_detector_position that returns the
position of the center of the detector bank, that is used in ISISCommandInterface

re #5942