Refs #6339. Fixed several tasks. One is left
The problem with writing and reading the DetCal file for new TOPAZ runs may
be fixed by others.

Rotations now rotate the center of the panel, not only the orientation of the
panel.

The algorithm now removes previous output tables so they do not clash( with a
LOT of warning pop ups) with the new output table.

Groups for grouping can now have commas separating the groups.

The optimizing for the sample offset still has to be done