Refs #6352. Getting peak coordinates to switch when asked.
I had to do this as a post-accept check to stop PV from setting the parameter
back to the default when I tried to set the parameter before the accept call.
Found out in the process that the signal connections in MdViewerWidget for
the MultiSliceView don't work anymore. That's why the check signature is now
different. NOTE: Things are currently wired to switch to sample coords. There
is sample code in SplatterPlotView::checkPeakCoordinates() to get things the
right way.