Remove CFBundleIdentifier from Info.plist Refs #6372
It caused issues with the installer trying to "upgrade" the build copy of
the application rather than installing it. Removing the identifier
means nothing matches so an upgrade is never attempted. (The real fix is
to pass --no-relocate to packagemaker to turn off auto relocation but
that's out of our hands as cpack doesn't allow the call to be customized)