RE 6397  Scripts for TOPAZ/MANDI SCD reduction
These scripts use Mantid, but are intended to be run
separately.  In particular, the ReduceSCD_Parallel.py
script includes a simple mechanism for reducing multiple
runs in parallel, by launching separate command line
processes for each run.  Each of these process is an
instance of ReduceOneSCD_Run.py.  The same .config
file can be used for reducing one run at a time with
ReduceOneSCD_Run.py and for reducing multiple runs
in parallel with ReduceSCD_Parallel.py.  ReduceSCD.config
is a sample configuration file.  It will have to be
edited to place the output files in the correct directory
and to reduce different runs with different control
parameters.  The python files and .config file are
heavily documented, so their use should be clear.
Typically, users should copy these files to their
home directories and customize them as needed.
refs #6397