RE 6476 Improve documentation for IntegrateEllipsoids
Added introductory "Overview" section.  Added "Explanation of
Inputs" section with diagram.  The original documentation is
now in a section titled "Detailed Algorithm Description".  The
equations for background subtraction are the same as for the
IntegratePeaksMD, so the user is referred to the documentation
for that algorithm, rather than cutting and pasting that part
of the docs to this algorithm.

refs #6476