Solve the problem of corrupted repository
If the user (could not wait for the installation) and canceled
the installation process, the folder for the scriptRepository may
be created, but the references and indexes are not valid. Further
attempt to use this repository may cause a segmentation fault.

Checking if the repository is empty help to prevent such case.

re #6175