Re #5174. Up memory limit tested against and add message to test.
I think this test still has some value in partly checking complete
nonsense isn't going into the variable. So I've increased the RAM
tested against to a value that isn't likely to be encountered
in the foreseeable future and added a failure message so that it's
clear what is being tested against.