Re #6606. Remove unsafe V3D constructor from a pointer.
Although it checked for a null pointer being passed in, it didn't
initialise the vector members in that case. Even if that were
fixed is would be unsafe as an array of the wrong length could
be passed in. It was only used in a few places, so I've elected
to just remove it.

This should fix the failing test on the ORNL mac build.