Definition of the unit test for ScriptRepository
PythonInterface: update the python interface that is exposed.
ScriptRepositoryImpl.h: set doDownloadFile as virtual, otherwise it is not override by the test mock.
scriptrepositoryparser : change the time format.
ScriptRepositoryImpl.cpp : solve some issues that appeared with the unit test.
ScriptRepositoryTestImpl: define many tests to ensure the functioning of the ScriptRepository.

re #6175