Add code to validate requested start time for live listener
Added code to LoadLiveData.cpp to validate the time string for the
requested start time.  Also use the epoch time for the "start from
now" option (instead of using the current time).  The epoch time gets
converted to 0 when passed to SMSD and SMSD handles that a little more
efficiently than passing it the current time.

Refs #6661