Solve the issue of wrong monitors numbers
Now the monitors are taken from the workspace, so, to be robust against
changes in these numbers. The spectra are set with the new api, that set
all the values using the numpy array. (Faster and easier)

Restore the old Load method, does not skip loading the event files any more.

re #6594