Re #6768. Expose AlgorithmProperty to python.
This is needed to avoid an error if you try to capture the return value
from StartLiveData in python, now that this algorithm has an output
property of this type.
You can't really do much with it right now, but if needed later
we could expose the IAlgorithm::cancel method.