Re #6811. Use AlgorithmManager's newestInstanceOf method.
In place of pulling out the list via the (now gone) algorithms()
method and going through it up at this level.
There is a slight difference in behaviour, in that the old way returned
the newest algorithm that HADN'T been executed already. However,
since the dialog is created immediately after the algorithm has been,
I don't see how that can be a problem.